requirements.txt
openai
chromadb
streamlit
tavily-python
beautifulsoup4
requests
python-dotenv